refactor: revamp searchers
Introduce AbstractSearcher to reduce duplicate logics in searchers
