+++ uzen/services/rule_matcher.py
@@ -5,7 +5,7 @@ import itertools
 from uzen.models.rules import Rule
 from uzen.models.schemas.matches import MatchResult
 from uzen.models.snapshots import Snapshot
-from uzen.services.rule_search import RuleSearcher
+from uzen.services.searchers.rules import RuleSearcher
 from uzen.services.yara_scanner import YaraScanner
 
 CHUNK_SIZE = 100
